requirements.txt
tensorflow>=2.15
tensorflow-addons
huggingface_hub
datasets
Pillow
pyyaml
sentencepiece